Add test to very entries in deps_info.py.
This test create a small test program for each key in deps_info.py.
The it checks the compiled JS output for usage of each of the
symbols declared as dependencies.  If a given symbol is not used then
the dependency deemed not to be invalid.

Adding this test means that as we evolve the code it won't ve possible
to accidentally forget to update this file.

Checking for symbols that are "used" is somewhat tricky.  Each symbol
will be defintition appear in the output JS since we are forcing the
export that those symbols.  We also don't want to match of substrings
or comments.